# Conflicts:
#	main-service/src/main/java/ru/practicum/mainservice/repository/RequestRepository.java
#	main-service/src/main/java/ru/practicum/mainservice/service/impl/EventServiceImpl.java

+++ main-service/src/main/java/ru/practicum/mainservice/model/Event.java
@@ -62,7 +62,7 @@ public class Event {
     @Getter
     @Setter
     @Transient
-    private int confirmedRequest;
+    private long confirmedRequest;
 
     @Getter
     @Setter
